refactor(be/pkg): Moving pkg dir to right place
